requirements.txt
Flask==3.1.0
torch==2.5.1
torchvision==0.20.1
Pillow==9.5.0
numpy==1.26.4